+++ jacob/jni/Variant.cpp
@@ -190,7 +190,8 @@ JNIEXPORT jobject JNICALL Java_com_jacob_com_Variant_toEnumVariant
     // SF-1674179 fix EnumVariants memory leak
     // AJ: yes, but the QueryInterface call above already incremented the reference
     //if (ie) ie->AddRef();
-    jobject newAuto = env->NewObject(autoClass, autoCons, ie);
+    // jacobproject/bug/132
+    jobject newAuto = env->NewObject(autoClass, autoCons, (jlong)(uintptr_t)ie);
     return newAuto;
   }
   return NULL;
@@ -411,7 +412,8 @@ JNIEXPORT jobject JNICALL Java_com_jacob_com_Variant_toVariantDispatch
     IDispatch *disp = V_DISPATCH(v);
     // I am copying the pointer to java
     if (disp) disp->AddRef();
-    jobject newAuto = env->NewObject(autoClass, autoCons, disp);
+    // jacobproject/bug/132
+    jobject newAuto = env->NewObject(autoClass, autoCons, (jlong)(uintptr_t)disp);
     return newAuto;
   }
   return NULL;
